feat: improved performance and code readability
