+++ pallets/orders/src/interface.rs
@@ -33,5 +33,4 @@ pub trait OrderInterface<T: frame_system::Config> {
 		account_id: &T::AccountId,
 		escrow_key: &T::AccountId,
 	) -> Result<(), Self::Error>;
-	fn is_pending_order_ids_by_seller_exist(account_id: &T::AccountId) -> bool;
 }
